refactor: 미사용, 중복 코드 정리
