--- src/main/java/com/koa/sws/model/Role.java
@@ -1,6 +1,0 @@
-package com.koa.sws.model;
-
-public enum Role {
-    PUBLISHER,
-    SUBSCRIBER
-}
